compile

+++ include/openssl/sm3.h
@@ -1,9 +1,9 @@
 /*
  * Copyright 2017 The OpenSSL Project Authors. All Rights Reserved.
  * Copyright 2017 Ribose Inc. All Rights Reserved.
- * 
+ *
  * Copyright zhizhi-dev@gmail.com. All Rights Reserved.
- * 
+ *
  * Licensed under the OpenSSL license (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
  * in the file LICENSE in the source distribution or at
@@ -28,7 +28,7 @@ typedef struct sm3_state_st {
   uint32_t h[8];
   uint32_t Nl, Nh;
   uint8_t data[SM3_BLOCK_SIZE];
-  size_t num;
+  unsigned num;
 } SM3_CTX;
 
 OPENSSL_EXPORT uint8_t *SM3(const uint8_t *data, size_t len, uint8_t out[SM3_DIGEST_LENGTH]);
